backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.2
bcrypt>=4.0.0
python-jose[cryptography]==3.3.0
celery==5.3.6
redis==5.0.1
sentence-transformers==2.5.1
faiss-cpu==1.8.0
langchain==0.1.12
openai==1.14.0
ragas==0.1.5
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
gitpython==3.1.42
tiktoken==0.6.0
pypdf==4.1.0
python-multipart
certifi
